refactor constructor

+++ nigbms/modules/meta_solvers.py
@@ -76,7 +76,7 @@ class Poisson1DMetaSolver(MetaSolver):
                 raise ValueError(f"Feature {k} not found in task")
 
         features = torch.cat(features, dim=1).squeeze()  # (bs, dim)
-        features = features.float()  # neural network expects floats
+        # features = features.float()  # neural network expects floats
         return features
